Title: 배열의 특정 위치 찾기, Time: 130ms, Memory: 10MB, Status: Passed - Codetree

+++ 250813/배열의 특정 위치 찾기/find-specific-location-fo-array.java
@@ -27,7 +27,7 @@ public class Main {
             cnt++;
         }
 
-        double avg = second_case_sum / cnt;
+        double avg = (double) second_case_sum / cnt;
 
         System.out.printf("%d %.1f", first_case_sum, avg);
